requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=9.0.0
scipy>=1.7.0
scikit-image>=0.19.0

# Data augmentation
albumentations>=1.3.0

# Visualization
matplotlib>=3.5.0
tqdm>=4.62.0

# SEEM model dependencies
detectron2
timm>=0.6.0
einops>=0.6.0
transformers>=4.25.0

# Utilities
pyyaml>=6.0
omegaconf>=2.3.0

# Optional: for development
# black>=22.0.0
# flake8>=4.0.0
# pytest>=7.0.0
